+++ src/ppf2.cpp
@@ -168,6 +168,6 @@ void ppf2(const Graph& g, Vertex first_right, VertexVector& mate, int numThreads
 #pragma omp parallel num_threads(nt)
 #pragma omp for
     // update matchings for left vertices
-    for (int y = first_right; y < n; y++) if (is_matched(y, mate)) mate[mate[y]] = y;
+    for (int y = first_right; y < n; y++) if (is_matched(y, mate)) mate[mate[y]] = static_cast<int>(y);
 }
